+++ NEWS.md
@@ -1,5 +1,11 @@
-# User visible changes in `FITSCards` package
+# User visible changes in `FITSBase` package
+
+## Version 0.2.0
+
+- Package renamed `FITSBase.jl`.
 
 ## Version 0.1.1
 
+- Package name `FITSCards.jl`.
+- Add `FITSHeader` structure.
 - New card properties: `card.units` and `card.unitless`.